Fixed a couple small problems with Linux

+++ sources/code/Engine/Core/EBase.h
@@ -4,7 +4,7 @@
 
 class EBase {
 protected:
-	size_t id;
+	unsigned id;
 public:
 	EBase();
 	unsigned int components[NUM_COMPONENTS];
